requirements.txt
# Discord Bot
discord.py>=2.3.2
aiohttp>=3.9.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23
alembic>=1.12.1

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# AWS Lambda (future)
boto3>=1.34.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
tqdm>=4.66.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0

# Optional caching
redis>=5.0.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
flake8>=6.1.0

